[app] Add link handling, which involved a bug fix
Only handles https://lobste.rs and https://lobste.rs/s/.* links for now

Bug fix:

The FrontPageDataSource is given a scope to do coroutine stuff in.
But since my components are all created and cached in ClawApplication,
and the scope was given in the FrontPageModule, the scope was
essentially a singleton, which is wrong, especially when multiple
activities start popping up like when handling links.  So, embrace the
Java, and make a FactoryFactory