[feature-comments] WIP Start making a custom link movement method
This will allow me to replace URLSpans with a pressable URL span for
prettiness.  Also, take away text selectability, add it back later
as a long press option. Als, need callbacks in the movement method to
let the view holder handle clicks, including a callback for when the
text view is action_downed but not on a span, which will let the view
holder (or probably the adapter -> controller) know if it should
collapse predecessors or self