Render the About section in User Profile as markdown
Using the commonmark library.  Works like a charm,
Markdown -> Html -> Spans, looks just like the comments.
Might not be the most effecient but it's usually less than a screenful

Also changes the text gravity of the about so it doesn't look like garbage,
but still moves to the right if there is space